Client and unique ids splitted
! Breaking change !
ClientIDMode default value is Static, DefaultBehaviour was removed. When in Repeater or GridView (where DataItemContainer is used) ClientIDMode is AutoGenerated.

Dotvvm does not use id attribute any more, unique ids are generated just from the UniqueId property from compilation and data-dotvvm-id is rendered when id is needed on client side (it is actualy needed only for Postback.Update)

In Static mode the ClientID (html id attribute) is equal to the ID in markup. AutoGenerated is like static until the control is in a naming container. Then the id will be prefixed by the naming container's ClientID, ClientID of it's first ancestor which has Id and is not naming container or NC's unique id.

Btw. html attribute bindings are now rendered in server rendering mode.