Redesigned resource link controls:
You can now explicitly say "I want to have the resources rendered here" using <dot:HeadResourceLinks /> or <dot:BodyResourceLinks />. The resources are rendered on the first place where this control occurs.
HtmlGenericControl renders this control implicitly if the TagName is body or head (I have moved the logic from DefaultOutputRenderer, I think that it should improve performance)
ResourceManager now knows if the resources have been already rendered and it will throw an exception when a new one is added. This is potentially a runtime breaking change, but it should only throw on code that is actually incorrect.