Implemented support for delegate arguments in bindings
All bindings now have an implicit conversion to delegate - essentialy, when the expected type is a delegate, the binding is wrapped in lambda. Now it works in value and resource binding, in Javascript it's representer as a function.

When the delegate contains any parameters, these get registered as addition symbols to the parser - these are parameter with special ExtensionParameter which can't be translated into JS nor evaluated on server, but it's removed by the lambda implicit conversion.

To use the parameters with staticCommand you can set the commandArgs symbolic parameter (the same way you'd do with command binding)

see #527 for context