inner element properties tweaks
InnerElement properties are not limited to elements - now, it can also contain literals, ...
When the default property is empty or whitespace it's not assigned (it's null). However, when the name is specified, the whitespace content is included (when valid in this context).
The content is filtered by the specified collection type (e.g. GridViewColumn), error is reported only when the element is not asignable to the specified type and when it's not empty.