Redesigned how _index works
Removed JavascriptTranslator.CurrentIndex symbolic parameter, it's usage
is replaced with `$context.$index` (as we'd do with any other
non-special property of the knockout context). This fixes a bunch of
issues with data context shifting of the context.

Repeater and GridView in server rendering mode use dotvvm-SSR-foreach
and dotvvm-SSR-item binding handlers that:
1. Automatically assign `$index` in each template
2. Remove the DataSource expression from every item of the Repeated

Method BindingHelper.SetDataContextForItem was removed since it can't be
implemented without the CurrentIndexParameter in the ParametrizedCode.

This commit should fix issues with $index in server rendering, $index in
staticCommand and $index in somewhat deeply nested data contexts.