Fixed unique ID generation for some cases
We are now using unique id counter instead of relying on List.Count,
which solves problems with controls added in a strange order (like
`x.Add(controlA); x.Add(controlB); x.Remove(controlA); x.Add(controlC)`,
where B and C remain in `x` with the same id).

When a control is added to the collection, it's id is assigned only if
it's parent also has the id. And when the id is assigned, it's assigned
recursively to the entire subtree that does not have any ids. This solves
problems with entire hiearchies being added into the control tree (that
are heavily used in DotVVM BP)