Disabled nested and unsupported bindings in GetValue
This simple change disables evaluation of bindings returned by binding.
This was supported for very long time, but never had any practical usage
(as far as I know). With DotVVM 2.0 is became quite impractical to have
a binding that would return another binding, since it should also
declare it's return type. The binding also should be translatable to JS
and in that case it can not return another binding at all.

This "feature" now only creates quite unpredictable runtime decision
about the result type and I think we can safely remove it.

Another problem arises when you store something that is actually not
evaluatable - an IBinding instance that is not ICommandBinding nor
IStaticValueBinding. In that case, it simply hang in an infinite loop,
which is clearly a bug. With this change, it will throw an exception,
it's safer that silently returning it and it will not break any
currently working scenario.