Added API for runtime warnings
RuntimeWarningCollector is a service, where a new warning will can be
added using the `Warn` method. Warning is an instance of
DotvvmRuntimeWarning, contains a message and may contain a reference to
an exception and a bindable object.

By default, warnings are collected in Debug mode. They are displayed in
browser JS console and forwarded to Asp.Net Core logging.

IDotvvmWarningSink is extensibility point for people that want to
collect these warnings on their own.

As a demo, HtmlWriter reports warnings when elements that should be
always self-closing contain some content, as the browser will interpret
the content as sibling