Implemented LRU Cache and used it as default
This commit should solve the problems with the default caching mechanism (with adapters):
* on OWIN it was completely broken, so I have removed it completely
* on Asp.Net Core it only had problems with cache retention, so I have only changed the default
    - in general, the MemoryCache is not a great fit for our use-case since we don't know the binding size,
      we don't have any expiration and we don't know how much memory can we use.

The new LRU cache will have a predefined size and predefined retention interval and will eject the items
that were unused for longest time (not created first as expiration time would do).

The default cache size is 1000...2000 (it should at most 5MB if it's full of bindings)
and the default timeout is 10 minutes.

Note that I'm not sure whether we want to use LRU dictionary as implementation of the Cache adapter
or rather use it strictly for the binding cache. In that case, however, the adapter mechanism becomes unused.