Generate unique ids for markup control viewmodules
we use the control unique id.

The id also has to be used in any commands invoking the module jsCommands. We can however reuse the infra for standard commands that also require control ID.
Only issue could be when a command would migrate from one markup control to another,
but that's considered unlikely given we add support for invoking _control.Command from staticCommands