Fix registration of modules in markup controls
Modules should be registred only for the page which have the @js directive.
The property controlling this behavior was inherited, which caused this.

Incidentally, this also fixes part of #941 - when the markup control is used
from a page. The issue still persits when the templated control is used from a
markup control.