Add a MSBuild task for running `yarn build`
In case yarn is not installed, npm is used.
Yarn is preferred, because it takes less time to run it, so it's
less annoying.
The Target only runs when any typescript file is changed,
otherwise it's skipped by MSBuild

In case there are some JS error, the build should continue normally,
so hopefully it won't cause some build issues

This should solve the problem that we forget to build the JS and
then something does not work. And when someone clones the repo,
it just won't compile because the JS files are missing