Remove support of virtual properties
resolves #802

They can be removed since capabilities now do a better job without
all the glitches.

Main reason for removal is that it leads to very unintuitive behavior when
the property is not registered for some reason (usually explicit assembly loading)

It also created redundant edge cases on some places and it's not really used in practise.
The error message hopefully explains what is going on.