static commands: fix parameter passing (in async commands)
The parameters were passed in a very weird way, now it works similar
to command bindings - even when command is not a lambda function, it is
wrapped in a lambda function. Then, the lambda function gets parameters
assigned directly, so the lambda syntax is essentially eliminated.
This back and forth is kinda needed to align how it works in case command
is specified as plain expression and when it's specified as
lambda function.

resolves  #980, #1165