Add Save & Share button to error page, redact cookies
Since we don't know on server which cookies are HttpOnly, we
don't render anything (only the cookie names).
I have added a piece of JS code which fills in the cookies which
can be found in document.cookie

Added button that saves the error page as one Html file, so it can be
easily shared.