Fix isViewModelUpdating with ko.delaySync
The problem is that we used to set isViewModelUpdating to false before
we invoked ko.delaySync.resume(). This lead to actually updating the controls
after when isViewModelUpdating was already false.

I also moved invocation of the newState event into this isViewModelUpdating=true
block, since it's also used for updating controls