Use http security headers by default
Switches were added to DotvvmSecurityConfiguration, so the behavior
can be configured.

* X-Frame-Options: DENY is used by default. It should pose no problem,
 since DotVVM did not work with iframes anyway due to cookie
  SameSite policy. When CrossSite frames are enabled, the cookie will
  have SameSite=None
* we also check Sec-Fetch-Dest - if it's an iframe, we validate on server that iframes are allowed.
   This mostly allows us to print a nicer error message, but may also in theory prevent some iframe timing attacks
* X-XSS-Protection: 1; mode=block
   probably not super useful, but also not harmful
* X-Content-Type-Options: nosniff
   - disable inference of content-type based on content, this could prevent some XSS attacks.
   probably also not super useful, but very unlikely to cause any problems.

* When on https, we set the session cookie with __Host- prefix. This prevents
 it being used by subdomains. Can help only in obscure cases -
 but for example when a system on the subdomain is compromised,
 the attacked can not pivot to the parent domain so easily.

We also check Sec-Fetch-* headers - that tells us what the browsers intends
to do with the page and if it's cross-origin or same-origin request.
Basically, we don't allow cross-origin POST and SPA requests. We also
don't allow JS initiated GET requests to dotvvm pages, but this may
be disabled, as someone could rely on it.

If the browser does not send these Sec-Fetch-* headers, we don't check
anything. You can enable strict checking by Security.RequireSecFetchHeaders option.
By default it's enabled on compilation page to prevent SSRF
and it does not matter too much if it does not work for someone.

Importantly, we don't set Strict-Transport-Security, since it's quite hard
to undo if it's unwanted and should be set on all requests coming from the server,
not only from requests handled by dotvvm.

I basically went through this guide and implemented what makes sense in our context:
https://infosec.mozilla.org/guidelines/web_security#web-security-cheat-sheet