Add basic support for capabilities in server-side styles
Just basic support to be able to at least get/set properties
only available through capabilities. The API is fairly simple -
it's just an extension of the existing parser from Linq.Expression:
`GetProperty(c => c.Prop)`. Now it allows more complex expressions:

* c => c.Capability.Property
* c => c.PropertyGroup["constant"]
* c => c.PropertyGroup[localVariable]
* c => c.Capability.PropertyGroup["x"]

Note that the same parser is used for the control.SetProperty(..., ...) which
available at runtime, so I tried to not make it very slow.
It might be convenient to use it there as well, but it's not
recommended in performance sensitive scenarios (like in dotvvm controls)