markup controls: Don't render unused properties
Another crack at reducing the amount of bullshit HTML we produce...
Just added a visitor which goes through all client-side bindings
and checks which properties are actually used.
Then, when rendering a markup control, only those will be included.

It also reduces the possibility of invoking arbitrary commands bound to
the markup control.
And it will stop exposing objects which may not be intended for the user.

It is technically a breaking change for people accessing the $control field
manually, but I hope there is not that many of them.