Make it possible to access entire capability in server-side styles
For example, this makes it possible to write a common function for a style
and for a control.

Potentially, this infrastucture could be used to execute
entire composite controls at compile-time. This makes is a bit more
limiting, since it's only executed once, but if you wish
to create data bindings in the control, it's the only option
that will not make the page super slow just by putting
the control in a Repeater