Add Referrer-Policy: no-referrer by default
This should help protect user privacy and security.

Two options are added, one to disable this feature completely and this
can be used for specific routes that need the Referer header for some
reason. Another option controls the header value - it may be reasonable
to set it to same-origin to allow basic user tracking.

We have it set to no-referrer by default given that we want to use
the strictest feasible measures - the developer always has the option
to turn it off. Also, this is quite unlikely to break anything apart
from basic tracking tools (which is arguably a good thing :] ...
I should also figure out how to break the more advanced trackers).