Add _js.InvokeAsync method
Before it was used as Invoke<Task<X>>, which is pretty ugly.
This patch also adds a check that synchronously invoked command
does not return a promise, otherwise it would probably fail with
some undefined access error.
Checking that an async command actually return a promise is unnecessary,
since await in JS is a no-op
for synchronously evaluated values.