Fix continuous integration (GitHub Actions) (#1234)
Fixes and extends our CI pipeline on GitHub.

**Jobs:**

_Build all projects without errors:_
Builds truly **all** projects and fails if compilation fails. This should prevent us pro merging commits that mess up compilation of our more peripheral projects (or those that only run on Windows). My reasoning behind this addition is that we don't want people to come to this repo and clone it only to find out that the solution cannot be compiled.

_Build published projects without warnings:_
Same job as what _build without warnings_ used to be, but builds all projects that we publish on NuGet. They are listed individually in the action, so that it's obvious which failed the job.

_.NET unit tests:_
Runs `src/Tests` and `src/Analyzers/Analyzers.Tests` and creates reports if there are any failed tests.

_JS unit tests:_
No changes were made to this job. Still runs our JS tests.

_UI tests:_
Runs UI tests on both Chrome and Firefox and both ASP.NET Core and OWIN. Creates reports if there are any failed tests.

**Composite actions**

* `.github/setup` -- sets up .NET and Node, restores and caches packages, and adds a `Build-Dotvvm` script that serves as shortcut for
  ```
  "msbuild.exe" -t:Clean,Build -v:m -warnAsError -p:Configuration=$configuration -p:TargetFramework=$targetFramework "$project"
  ```
* `.github/uitest` -- runs our UI tests on a given browser and publishes the results. If the runner's OS is Windows, runs the OWIN version.
* `.github/unittest` -- runs a given test project and publishes the results.
* `test-report` -- reports **failed** tests found in a `.trx` file using GitHub's [checks API](https://docs.github.com/en/rest/reference/checks). Reports nothing if all tests passed or skipped.

**Other changes**

* Replaces npm with yarn during CI.
* Converts old-style project to the new csproj style with [MSBuild.SDK.SystemWeb](https://github.com/CZEMacLeod/MSBuild.SDK.SystemWeb).
* Regenerates assembly redirects automatically (hence the changes to `web.config`s).
* Removes configuration for our **public** Azure Pipelines CI.

**Outstanding issues**

* Tests running in Firefox on Windows are insanely slow. This seems to be an issue with SeleniumUtils.
* There still isn't a deployment pipeline on GA
* Currently both .NET 5 and .NET 6 need to be set up. #1272 seems to fix this.