Incomplete type members no longer participate in serializability analysis
This is a small adjustment to improve situation where we immediately report diagnostics on unfinished type members. For example user tries to define a property `public int Prop`, at this point they might get a warning that fields are not supported.

The purpose of this patch is to analyze only type members that do not have compilation errors attached (CSxxxx)