ToSystemType will try finding the type without assembly name
This caused problems for VS extension where it was provided
with `System.Boolean, netstandard`, but we would need to
use mscorlib for it to be found. Way around this is that
core types can be found without the assembly name,
which allows us to not rely on specific core assembly name.

Also added TryToSystemType which won't throw an exception