Repeater calls templates with rooted parent
Usually it's not super important if the container control is rooted or not,
but when the DelegateTemplate contains some user code which calls
GetDotvvmUniqueId (or similar), it will fail in weird ways (the ids will be
generated but won't be unique). For components used only from markup,
this really does not matter, but Repeater is being used as a base component
for many composite controls.