Add AddCssClass and AddCssStyle extension methods for ValueOrBinding
Some not very nice code has to be used in Fluent UI and BS5 for this,
so it would be nice to have this helper.

In order for AddCssClass to work with any html control, not only
HtmlGenericControl, generic mechanism for handling property
groups from capabilies was introduced
 - see DotvvmCapabilityProperty.FindProperty
 - and see DotvvmBindableObjectHelper.CapabilityCache.cs

the FindProperty can lookup any property from a specified capability.
The CapabilityCache makes it reasonably fast specifically for
HtmlCapability