Fix late initialized template in JsComponent
When a KnockoutTemplateReactComponent was added to
DOM after the initial render, its element woudn't have the
knockout context, so it couldn't be initialized. The fix is to
simply get the knockout context from the nearest parent
element.

This also allows us to not apply knockout bindings to the
React/Svelte generated elements, and we can get rid of
the hack with setTimeout

Thanks @lorencjan for reporting the issue and suggesting
the way to fix it