hierarchy repeater: fix client-side command in markup control
problem was that that BindingHelper.FindDataContextTarget
assumed there was one more data context layer since we used set
data context type and DataContext on different controls.
Because of that command client-side and server-side
data context paths didn't match.
Technically, this is a glitch of BindingHelper, not HierarchyRepeater,
but "fixing" BindingHelper will break many things and it's hopefully
going to be done in #1392