Extend DotvvmFeatureFlag API, improve validation
The API was originally designed for features which are disabled by default,
but it's now being used for features enabled by default (mainly security headers).

For this reason I'm adding the DisableForAllRoutes method. DisableForRoutes is not neccesary,
EnableForAllRoutesExcept has exactly this semantic. The name might be a bit confusing in
this context, but I'm not sure adding a synonym will help.

Instead I'm adding methods for incrementaly including/excluding additional routes:
* IncludeRoute(string) / IncludeRoutes(string[])
* ExcludeRoute(string) / ExcludeRoutes(string[])

These new methods can be used in fluent style to save some keystrokes, for example

config.Security.FrameOptionsCrossOrigin
    .DisableForAllRoutes()
    .IncludeRoute("route1")
    .IncludeRoute("route2");