Render html attribute values even in client-rendering mode
In all cases, bindings are evaluated in a try { } block, so
no error should be raised. Technically, adding an attribute
which wasn't there is a breaking change, but since
the attribute will be replaced client-side, it's unlikely to
matter. On the other hand, especially certain attributes
(like href, alt, title, class) are rather important to show
ASAP or show to no-JS indexers