AutoUI: nested contexts should contain _index and _collection
We had slight mismatches in the data context types of
selector bindings - the control context normally includes
_index and _collection extension parameters, but the
bindings created by AutoUI did not. This didn't cause
any issues, since ComboBox doesn't check it and
the multi selector contained a DataContextType override.

However, if the same logic was used as is in other places (BP),
it was crashing with data context type mismatch exception.