Redesign comments in generated Javascript
The JsCommentNode approach is problematic - either we need
to enable formatting any children in any nodes, or special
case the comments in the formatter.
Another approach would be to make the node an
expression which could wrap any other expression, but
that complicates all code peeking into the JsNode.

This patch adds a CommentBefore property to all JsNodes,
and adds the code to format these comments into the JsFormattingVisitor.
Although the formatting visitor has to check for comments on each node,
no other code has to worry about weird nodes in the tree.