Run all JsComponent interactions asynchronously
Initialization, update and dispose of JsComponent
is now executed in the next event loop cycle - i.e.
after all other knockout components have been initialized.
The main reason for this is to NOT handle
exceptions from the JS component and avoid
crashing the entire page when one component fails to load.