More reasonable unclosed tag handling
Enabling the warnings in #1659 reveals that our handling of unclosed
tags is not ideal. This patch is attempt to improve it, while maintaining
reasonable level of backwards compatibility.

The only behavior change is what happens when </x> has no opening tag
* previous: close currently open elements, then insert <x />
* new: insert <x /> into the current element

a warning is issued in both cases. The new logic avoids a major problem
of one accidental end tag closing the entire hierarchy, including
the <body> and <html> element. Since we insert script references at
the end of body, this puts the scripts and viewmodel wherever the closing tag was.
This has the potential to break something.

Other changes only affect the issued warning:
* no warning on an unclosed tags like img, meta, br (which are always self-closing in HTML)
* no warning on automatically closed tags like p, td. In HTML it's
   perfectly valid let these tags close automatically, so we can also
   consider this valid.