Support PostBack.Handlers in CompositeControls
PostBack.Handlers are copied onto the created controls, if they
contain a matching command binding.

* We map postback handlers onto the command bindings set on
   the CompositeControl by comparing the EventName and property name
* For each control returned from GetContents, we enumerate its
   command bindings and compare them to the command
   bindings found on the CompositeControl.
* If a matching command binding is found, we clone its postback
   handlers onto the child control. EventName is adjusted to
   match the new property name

The control enumeration is done recursively before adding the
control to Children, thus walking only through the tree created in
this CompositeControl - nested CompositeControls,
initialized Repeaters, ... are not included.

In order to support templatws created in the CompositeControl,
we recurse into CloneTemplates. For DelegateTemplate or any
other more advanced needs, a protected
CopyPostBackHandlersRecursive method is provided.

resolves #1699