CompositeControl: better handling of ambiguous GetContents, test for inheritance
According to  #1613 CompositeControl do not work with
both base and derived types defining GetContents method.
I could not replicate that, the provided example works for me,
maybe it was fixed accidentaly.

* I added test for that
* In any case, multiple GetContents overloads were handled poorly,
   I added better error for this case.

In the issue, I mentioned that it would be complex to handle
potential redefinition or "removal" of properties in the derived type.
That isn't really the case, both cases don't need any special code
* The "removal" - omission of an property in the derived type is simply ignored.
   I think that it's quite intuitive that properties can't disapear in
   derived types, if the GetContents doesn't ask for it it won't get it.
* Redefinition is already handled reasonably, since we can already
   try to redefine IncludeInPage or some capability property.

Resolves #1613