Allow weird assembly name
I didn't find and documented restrictions on the assembly names.
If we take the new AssemblyName(...) function as reference implementation,
basically only restrictions are no `,`, `=`, `/`, `\` and nullbytes
(invalid UTF-16 is also allowed 🤦 :D)

So this allows whatever we can allow. Assembly name is always at the end,
so there is no point in placing restrictions on it.

resolves  #1728