Add back support for manually handling POST requests
We used to ignore the request method and only decide if we should run
a command or a page initial "GET" based on the headers. This got rewritten
in 4.2 and thus broke anyone manually doing and handling POST requests.
It has been "expoited" by some users to do "API requests", but can
also be used to implement classic POST forms,
as a no-JS or shitty network fallback.

To implement a no-JS form in DotVVM with this patch:

    <form method="POST">
        <input type="text" name="something" />
        <button type=submit name=submit value=my-form>Submit</button>
    </form>

and handle it in view model:

    public override async Task Load()
    {
        var req = Context.HttpContext.Request;
        if (req.Method == "POST")
        {
            using var body = new StreamReader(req.Body);
            var data = HttpUtility.ParseQueryString(await body.ReadToEndAsync());
            var submitedForm = data["submit"];
            var something = data["something"];
        }
    }