New dot:ModalDialog control, wrapper for <dialog>
The control closes or shows the modal based on an Open property.
The modal is always shown using the .showModal() method,
non-modal dialog is already accessible by binding the open
attribute of dialog HTML element.
The Open property may either be a boolean or a nullable object,
the dialog is shown if the value isn't false nor null.

On close event, false or null is written back into the Open property.
Otherwise, we'd quickly have inconsistent viewModel whenever
the user closes the dialog with ESC.
Close event is also provided for explicit event handling.

We also optionaly provide a helper for implementing
"close after backdrop click" functionality. It is not supported by
the dialog element natively and could not be otherwise
implemented without writing custom JS.
It is enabled by setting CloseOnBackdropClick=true

Resolves  #1708