Fix DotVVM02 warning on overriden ignored property
We now ignore override properties if the base has JsonIgnore
or Bind(None) attribute.

Technically, the patch introduces a false negative if the base
is ignored and the derived isn't. However, DotVVM behavior
regarding this is complex (you cannot override JsonIgnore,
only Bind(None)...) and given the rarity of virtual properties,
I don't want to unnecesarily complicate the analyzer.

fix #1784