capabilities: avoid NRE when ValueOrBinding<int>? is set to null
Nullable ValueOrBinding, for example ValueOrBinding<int>?
semantially means that the property is not set.
If the property is set to null, we would like to return null inside
of the ValueOrBinding.
However, that is not possible if T is non-nullable value type, so
either have to throw, or return an "outer" null. This patch
changes the behavior to return null instead of failing. Users
are unlikely to care about "not set" and "set to null", and
throwing the exception creates unnecessary friction.