Add info about pre-loading to the Sec-Fetch-Dest GET check
We check that the page is not being loaded with
JS, which should prevent attackers from scraping
all other pages if they gain access to one.

It seems that this check is incompatible with Chrome
prefetching (and obviously also any JS-based prefetching).
The issue seems to be that Chrome does not know
how will the prefetched document be used.

According to their documentation, it should ignore
prefetch requests with non-2XX status code.
Our rejections are 403, so it should not display them
to the user.