Maybe fix #1903: markup control property values are wrapped recursively with observables
These observables need to have .state and friends in order to work correctly
in staticCommands. In theory, the mapper should only process objects created
inline in the binding expression (such as in LINQ operators), meaning it's
unlikely to be a large object. Therefore, we just re-create the hierarchy on
each update, trying to re-use any subobject which is already wrapped
with `DotvvmObservable`s