Fix default values of PostbackScriptOptions
Resolves #1908

We now store all options as nullable, and the default value is determined by the KnockoutHelper function called.
It is necessary, as GenerateClientPostbackLambda has a different set of default values than GenerateClientPostbackLambda, and GenerateClientPostbackScript.

Notable exception is the ReturnValue, which already is nullable and null means that `null` should be returned.
It's thus technically difficult to make this property "optional", and this issue isn't really relevant for this property anyway