Disable lambda conversion in normal bindings
In command bindings, expressions get automatically wrapped in a lambda function. Surprisingly, the
behavior was also present in value/resource, which lead
to some surprising effects when value/resource was
accidentally assigned to a command property (#1742).

We also unify that untyped ICommandBinding and IStaticCommandBinding
produce delegate of type Command (i.e. async function with no argument and no return value).
Previously Delegate was used for commands,
which had the effect of sending the return value to the client