So we can't remove the if: dotvvm.isSpaRead condition
...but, we can replace it with `if: true`, which does not have the unwanted effect of hiding
the content before everything loads.

The reson is that ko.bindingContext/updateContext is pureComputed and won't update
$data right after we replace the viewModel. We need someone to call `context[contextSubscribable]` to update it.
Any binding will do that, but only after it's initialized, so the initialization will fail on
"Property X was not defined".

The pureComputed was introduced in https://github.com/knockout/knockout/commit/0ac7010001cd01748fd7bcfd948c06b2369d9bd7.
It's probably not a knockout bug, as it's not possible to reproduce without our hacks with
applyBindings in restoreUpdatedControls. If we used a binding handler for this, everything
would be fine.